Add trained model, ECS deployment, and Chrome extension
- Fine-tuned DistilBERT classifier (98.5% accuracy on stress test across Claude, GPT-4o, and real Yelp reviews)
- FastAPI backend deployed to AWS ECS Fargate via ECR
- Chrome extension (Manifest V3) for Yelp and TripAdvisor with background service worker for API calls
- Stress test results, training metrics, and model evaluation artifacts
- Updated README with architecture, deployment steps, limitations, and next steps

+++ scripts/generate_fake_reviews.py
@@ -236,7 +236,7 @@ def main(target_count: int):
     import pandas as pd
     df = pd.DataFrame(records)
     print("\nStar rating distribution:")
-    print(df["star_rating"].value_counts().sort_index().to_string())
+    print(df["star_rating"].astype(int).value_counts().sort_index().to_string())
     print("\nCuisine distribution (top 10):")
     print(df["cuisine"].value_counts().head(10).to_string())
